+++ appdaemon/plugins/hass/hassapi.py
@@ -264,40 +264,6 @@ class Hass(adbase.ADBase, adapi.ADAPI):
         self.call_service("persistent_notification/create", **kwargs)
 
     #
-    # Event
-    #
-
-    @hass_check
-    def fire_event(self, event, **kwargs):
-        self.AD.logging.log("DEBUG",
-                  "fire_event: {}, {}".format(event, kwargs))
-        
-        namespace = self._get_namespace(**kwargs)
-        if "namespace" in kwargs:
-            del kwargs["namespace"]
-            
-        config = self.AD.plugins.get_plugin(namespace).config        
-        if "cert_path" in config:
-            cert_path = config["cert_path"]
-        else:
-            cert_path = False
-
-        if "token" in config:
-            headers = {'Authorization': "Bearer {}".format(config["token"])}
-        elif "ha_key"  in config:
-            headers = {'x-ha-access': config["ha_key"]}
-        else:
-            headers = {}
-
-
-        apiurl = "{}/api/events/{}".format(config["ha_url"], event)
-        r = requests.post(
-            apiurl, headers=headers, json=kwargs, verify=cert_path
-        )
-        r.raise_for_status()
-        return r.json()
-
-    #
     # Service
     #
     @staticmethod